Revert "Move midi sending code to SoundDrum"
This reverts commit 68d86d5005f946b0623578c9a5b4a7049c096d69.
