+++ src/deluge/io/midi/midi_engine.h
@@ -25,7 +25,6 @@
 
 class MIDICable;
 class MIDIDrum;
-class SoundDrum;
 class MIDIInstrument;
 class MidiFollow;
 class PlaybackHandler;
@@ -43,12 +42,12 @@ struct MIDISource {
 	MIDISource(PlaybackHandler const* handler) : source_(handler) {};
 	MIDISource(MidiFollow const* follow) : source_(follow) {};
 	MIDISource(MIDIDrum const* drum) : source_(drum) {};
-	MIDISource(SoundDrum const* drum) : source_(drum) {};
+	MIDISource(Sound const* sound) : source_(sound) {};
 
 	MIDISource(MIDICable const& cable) : source_(&cable) {};
 	MIDISource(MIDIInstrument const& instrument) : source_(&instrument) {};
 	MIDISource(MIDIDrum const& drum) : source_(&drum) {};
-	MIDISource(SoundDrum const& drum) : source_(&drum) {};
+	MIDISource(Sound const& sound) : source_(&sound) {};
 	MIDISource(MidiFollow const& follow) : source_(&follow) {};
 	MIDISource(PlaybackHandler const& handler) : source_(&handler) {};
