+++ src/deluge/processing/sound/sound_instrument.cpp
@@ -402,24 +402,13 @@ void SoundInstrument::sendNote(ModelStackWithThreeMainThings* modelStack, bool i
 			}
 #endif
 			ModelStackWithSoundFlags* modelStackWithSoundFlags = modelStack->addSoundFlags();
-
-			noteOffPostArpeggiator(modelStackWithSoundFlags, instruction.noteCodeOffPostArp[n]);
+			if (allowNoteTails(modelStackWithSoundFlags, true)) {
+				noteOffPostArpeggiator(modelStackWithSoundFlags, instruction.noteCodeOffPostArp[n]);
+			}
 
 			reassessRenderSkippingStatus(modelStackWithSoundFlags);
 		}
 	}
-}
-
-void SoundInstrument::noteOnPostArpeggiator(ModelStackWithSoundFlags* modelStack, int32_t noteCodePreArp,
-                                            int32_t noteCodePostArp, int32_t velocity, int16_t const* mpeValues,
-                                            uint32_t sampleSyncLength, int32_t ticksLate, uint32_t samplesLate,
-                                            int32_t fromMIDIChannel) {
-	Sound::noteOnPostArpeggiator(modelStack, noteCodePreArp, noteCodePostArp, velocity, mpeValues, sampleSyncLength,
-	                             ticksLate, samplesLate, fromMIDIChannel);
-}
-
-void SoundInstrument::noteOffPostArpeggiator(ModelStackWithSoundFlags* modelStack, int32_t noteCode) {
-	Sound::noteOffPostArpeggiator(modelStack, noteCode);
 }
 
 ArpeggiatorSettings* SoundInstrument::getArpSettings(InstrumentClip* clip) {
